+++ mobile/android/android-components/components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/TabViewHolderTest.kt
@@ -28,7 +28,7 @@ class TabViewHolderTest : TestCase() {
             ) { /* noop */ }
         }
 
-        expectException(UnsupportedOperationException::class) {
+        expectException<UnsupportedOperationException> {
             simpleTabViewHolder.updateSelectedTabIndicator(true)
         }
     }