+++ mobile/android/android-components/components/concept/engine/src/test/java/mozilla/components/concept/engine/SettingsTest.kt
@@ -92,7 +92,7 @@ class SettingsTest {
 
     private fun expectUnsupportedSettingException(vararg blocks: () -> Unit) {
         blocks.forEach { block ->
-            expectException(UnsupportedSettingException::class, block)
+            expectException<UnsupportedSettingException>(block)
         }
     }