+++ mobile/android/android-components/components/lib/crash/src/test/java/mozilla/components/lib/crash/CrashReporterTest.kt
@@ -709,7 +709,7 @@ class CrashReporterTest {
 
     @Test
     fun `Internal reference is set after calling install`() {
-        expectException(IllegalStateException::class) {
+        expectException<IllegalStateException> {
             CrashReporter.requireInstance
         }
 
@@ -719,7 +719,7 @@ class CrashReporterTest {
             notificationsDelegate = mock(),
         )
 
-        expectException(IllegalStateException::class) {
+        expectException<IllegalStateException> {
             CrashReporter.requireInstance
         }